First try at skipping nodes without CIB connection
The sync marks based on pacemaker depends on the nodes to have a valid
CIB connection. This may not happen in case pacemaker is not up, the
node is fenced or in standby.

This proposal tries to skip the marks checks for those nodes that do not
have the CIB service active. Test show this implementation is not
working, but I think it can be the base for future work.