Added support for op defaults
We needed to add a default option for the pacemaker op defaults. In
particular we are interested in offering a default for "op monitor
on-fail" action.

Pacemaker does not offer a configuration option to offer such defaults
for all resources. There's a way to do it for a single resource with
rsc_defaults, but that is not our case. Also with Pacemaker you could
link one resource operations to another one with 'id-ref', but in that
case if you modify the reference resource it will cascade changes to all
other resourced.

This approach allows to store in the pacemaker barclamp a default for
op. In this case we are only interested in op monitor but could easily
be extended later on.

This change by itself does not fullfill the requirement. It needs the
'pacemaker-primitive' declared in crowbar-openstack repository to read
the value we store here. For that reason there's a dependency with a
change in that repository that creates a wrapper definition for
pacemaker_primitive that read the default from pacemaker barclamp.